Temperature Measuring Gun for the original code.

+++ User/inc/MLX90614_Driver.h
@@ -19,15 +19,15 @@
 
 void SMBus_StartBit(void);
 void SMBus_StopBit(void);
-void SMBus_SendBit(u8);
-u8 SMBus_SendByte(u8);
-u8 SMBus_ReceiveBit(void);
-u8 SMBus_ReceiveByte(u8);
-void SMBus_Delay(u16);
+void SMBus_SendBit(uint8_t);
+uint8_t SMBus_SendByte(uint16_t);
+uint8_t SMBus_ReceiveBit(void);
+uint8_t SMBus_ReceiveByte(uint8_t);
+void SMBus_Delay(uint16_t);
 void SMBus_Init(void);
 
-u16 SMBus_ReadMemory(u8, u8);
-u8 PEC_Calculation(u8*);
+uint16_t SMBus_ReadMemory(uint8_t, uint8_t);
+uint8_t PEC_Calculation(uint8_t*);
 float SMBus_ReadTemp(void); //获取温度值
